upgrade versions and rename to "meitrex"
