+++ .github/workflows/test.yaml
@@ -6,25 +6,32 @@ jobs:
   test:
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v3
+      - uses: actions/checkout@v4
         with:
           fetch-depth: 0
-      - name: Set up JDK 17
-        uses: actions/setup-java@v3
+      - name: Set up JDK 21
+        uses: actions/setup-java@v4
         with:
-          java-version: '17'
+          java-version: '21'
           distribution: 'temurin'
       - name: Validate Gradle wrapper
-        uses: gradle/wrapper-validation-action@ccb4328a959376b642e027874838f60f8e596de3
+        uses: gradle/wrapper-validation-action@v2
       - name: Run chmod to make gradlew executable
         run: chmod +x gradlew
 
-      - name: Execute tests
-        env:
-          SONAR_TOKEN: ${{ secrets.SONAR_TOKEN }}
-        uses: gradle/gradle-build-action@749f47bda3e44aa060e82d7b3ef7e40d953bd629
+      - name: Run tests
+        uses: gradle/gradle-build-action@v3
         with:
-          arguments: cleanTest test jacocoTestReport sonar
+          arguments: cleanTest test
+# Configuration for SonarCloud, enable this when SonarCloud is configured
+#      - name: Execute tests
+#        env:
+#          SONAR_TOKEN: ${{ secrets.SONAR_TOKEN }}
+#        uses: gradle/gradle-build-action@v3
+#        with:
+#          arguments: cleanTest test jacocoTestReport sonar
+
+
