+++ build.gradle
@@ -1,16 +1,16 @@
 plugins {
 	id 'java'
-	id 'org.springframework.boot' version '3.0.6'
-	id 'io.spring.dependency-management' version '1.1.0'
-	id "io.github.kobylynskyi.graphql.codegen" version "5.7.2"	
-	id "org.sonarqube" version "4.0.0.2929"
+	id 'org.springframework.boot' version '3.+'
+	id 'io.spring.dependency-management' version '1.+'
+	id "io.github.kobylynskyi.graphql.codegen" version "5.+"
+	id "org.sonarqube" version "4.+"
 	id "jacoco"
 	
 }
 
-group = 'de.unistuttgart.iste.gits'
+group = 'de.unistuttgart.iste.meitrex'
 version = '0.0.1-SNAPSHOT'
-sourceCompatibility = '17'
+sourceCompatibility = '21'
 
 def jacocoEnabled = System.properties.getProperty("jacocoEnabled") ?: "true"
 
@@ -34,7 +34,7 @@ graphqlCodegen {
 	// all config options:
 	// https://github.com/kobylynskyi/graphql-java-codegen/blob/main/docs/codegen-options.md
 	outputDir = new File("$buildDir/generated")
-	packageName = "de.unistuttgart.iste.gits.generated.dto"
+	packageName = "de.unistuttgart.iste.meitrex.generated.dto"
 	generatedAnnotation = "jakarta.annotation.Generated"
 	modelValidationAnnotation = "jakarta.validation.constraints.NotNull"
 	generateApis = false // set to false as the generator does not support spring boot graphQL
@@ -67,7 +67,7 @@ repositories {
 }
 
 dependencies {
-	implementation 'de.unistuttgart.iste.gits:gits-common:0.1.1'
+	implementation 'de.unistuttgart.iste.meitrex:meitrex-common:1.0.0'
 	implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
 	implementation 'org.springframework.boot:spring-boot-starter-graphql'
 	implementation 'org.springframework.boot:spring-boot-starter-validation'
@@ -80,14 +80,14 @@ dependencies {
 	runtimeOnly 'org.postgresql:postgresql'
 	annotationProcessor 'org.springframework.boot:spring-boot-configuration-processor'
 	annotationProcessor 'org.projectlombok:lombok'
-	testImplementation 'de.unistuttgart.iste.gits:gits-common-test:0.1.0'
+	testImplementation 'de.unistuttgart.iste.meitrex:meitrex-common-test:1.0.0'
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
 	testImplementation 'org.springframework:spring-webflux'
 	testImplementation 'org.springframework.graphql:spring-graphql-test'
 	testImplementation "org.mockito:mockito-core:3.+"
 	testImplementation 'org.hamcrest:hamcrest:2.+'
-	testImplementation "org.testcontainers:postgresql:1.18.3"
-	testImplementation "org.testcontainers:junit-jupiter:1.18.3"
+	testImplementation "org.testcontainers:postgresql:1.+"
+	testImplementation "org.testcontainers:junit-jupiter:1.+"
 }
 
 tasks.named('test') {
