+++ components/pubsub.yaml
@@ -1,7 +1,7 @@
 apiVersion: dapr.io/v1alpha1
 kind: Component
 metadata:
-  name: gits
+  name: meitrex
 spec:
   type: pubsub.redis
   version: v1
