+++ jacoco.gradle
@@ -1,5 +1,5 @@
 jacoco {
-    toolVersion = "0.8.8"
+    toolVersion = "0.8.11"
 }
 
 test {
@@ -9,8 +9,8 @@ test {
 
 jacocoTestReport {
     reports {
-        xml.enabled true
-        html.enabled true
+        xml.required.set(true)
+        html.required.set(true)
         html.outputLocation = layout.buildDirectory.dir('jacocoHtml')
     }
 }
