+++ settings.gradle
@@ -2,9 +2,9 @@ rootProject.name = 'template'
 
 sourceControl {
     gitRepository(uri('https://github.com/MEITREX/common')) {
-        producesModule('de.unistuttgart.iste.gits:gits-common')
+        producesModule('de.unistuttgart.iste.meitrex:meitrex-common')
     }
     gitRepository(uri('https://github.com/MEITREX/common_test')) {
-        producesModule('de.unistuttgart.iste.gits:gits-common-test')
+        producesModule('de.unistuttgart.iste.meitrex:meitrex-common-test')
     }
 }
