--- src/main/java/de/unistuttgart/iste/gits/template/config/package-info.java
@@ -1,4 +1,0 @@
-/**
- * This package should contain any classes that are used to configure the application.
- */
-package de.unistuttgart.iste.gits.template.config;
