--- src/main/java/de/unistuttgart/iste/gits/template/exception/package-info.java
@@ -1,6 +1,0 @@
-/**
- * This package is used for exception handling.
- * More info about exception handling with GraphQL can be found here:
- * https://www.baeldung.com/spring-graphql-error-handling
- */
-package de.unistuttgart.iste.gits.template.exception;
