--- src/main/java/de/unistuttgart/iste/gits/template/persistence/mapper/package-info.java
@@ -1,4 +1,0 @@
-/**
- * This package contains all the mapper classes that map between DTOs and entities.
- */
-package de.unistuttgart.iste.gits.template.persistence.mapper;
