--- src/main/java/de/unistuttgart/iste/gits/template/persistence/repository/TemplateRepository.java
@@ -1,14 +1,0 @@
-package de.unistuttgart.iste.gits.template.persistence.repository;
-
-import de.unistuttgart.iste.gits.template.persistence.entity.TemplateEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface TemplateRepository extends JpaRepository<TemplateEntity, Long> {
-
-    Optional<TemplateEntity> findByName(String name);
-
-}
