--- src/main/java/de/unistuttgart/iste/gits/template/service/package-info.java
@@ -1,5 +1,0 @@
-/**
- * The service classes implement the actual business logic. This usually includes the interaction with the
- * repositories and the mapping of the entities to DTOs.
- */
-package de.unistuttgart.iste.gits.template.service;
