--- src/main/java/de/unistuttgart/iste/gits/template/validation/TemplateValidator.java
@@ -1,12 +1,0 @@
-package de.unistuttgart.iste.gits.template.validation;
-
-import de.unistuttgart.iste.gits.generated.dto.Template;
-import org.springframework.stereotype.Component;
-
-@Component
-public class TemplateValidator {
-
-    public void validateTemplate(Template template) {
-        // add validation logic here
-    }
-}
