--- src/main/java/de/unistuttgart/iste/gits/template/validation/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Put class level validation logic in this package.
- */
-package de.unistuttgart.iste.gits.template.validation;
