+++ src/main/java/de/unistuttgart/iste/meitrex/template/config/ModelMapperConfiguration.java
@@ -1,0 +1,17 @@
+package de.unistuttgart.iste.meitrex.template.config;
+
+import org.modelmapper.ModelMapper;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+/**
+ * Provides a bean for the ModelMapper. The model mapper is used to map entities to DTOs and vice versa.
+ */
+@Configuration
+public class ModelMapperConfiguration {
+
+    @Bean
+    public ModelMapper modelMapper() {
+        return new ModelMapper();
+    }
+}
