+++ src/main/java/de/unistuttgart/iste/meitrex/template/config/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package should contain any classes that are used to configure the application.
+ */
+package de.unistuttgart.iste.meitrex.template.config;
