+++ src/main/java/de/unistuttgart/iste/meitrex/template/dapr/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package should contain all classes that are used to communicate with Dapr, e.g. using pub sub.
+ */
+package de.unistuttgart.iste.meitrex.template.dapr;
