+++ src/main/java/de/unistuttgart/iste/meitrex/template/exception/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * This package is used for exception handling.
+ * More info about exception handling with GraphQL can be found here:
+ * https://www.baeldung.com/spring-graphql-error-handling
+ */
+package de.unistuttgart.iste.meitrex.template.exception;
