+++ src/main/java/de/unistuttgart/iste/meitrex/template/persistence/entity/package-info.java
@@ -1,0 +1,13 @@
+/**
+ * This package contains the entities for the persistence layer.
+ * <p>
+ * Information on how to define Entities can be found in <a href=https://www.baeldung.com/jpa-entities>this tutorial</a>
+ * or <a href=https://docs.oracle.com/javaee/7/tutorial/persistence-intro.htm#BNBPY>this tutorial</a>
+ * or <a href=https://docs.spring.io/spring-data/jpa/docs/current/reference/html/#jpa.entities>the official documentation</a>.
+ * <p>
+ * Information on how to add validation annotations can be found in <a href=https://www.mastertheboss.com/java-ee/validation/test/>this tutorial</a>
+ * <p>
+ * To avoid having to implement getters, setters, builders, etc. we use <a href=https://projectlombok.org/>Lombok</a>.
+ * The Lombok annotations (e.g. {@code @Data}) should be used in the entities.
+ */
+package de.unistuttgart.iste.meitrex.template.persistence.entity;
