+++ src/main/java/de/unistuttgart/iste/meitrex/template/persistence/mapper/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package contains all the mapper classes that map between DTOs and entities.
+ */
+package de.unistuttgart.iste.meitrex.template.persistence.mapper;
