+++ src/main/java/de/unistuttgart/iste/meitrex/template/persistence/repository/TemplateRepository.java
@@ -1,0 +1,14 @@
+package de.unistuttgart.iste.meitrex.template.persistence.repository;
+
+import de.unistuttgart.iste.meitrex.template.persistence.entity.TemplateEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface TemplateRepository extends JpaRepository<TemplateEntity, Long> {
+
+    Optional<TemplateEntity> findByName(String name);
+
+}
