+++ src/main/java/de/unistuttgart/iste/meitrex/template/persistence/repository/package-info.java
@@ -1,0 +1,10 @@
+/**
+ * This package contains the repositories for the entities.
+ * <p>
+ * Note that we don't have to implement the interface, spring will do that for us.
+ * Simply extend from the JpaRepository and spring will provide the implementation.
+ * <p>
+ * For more information, see <a href=https://docs.spring.io/spring-data/data-commons/docs/1.6.1.RELEASE/reference/html/repositories.html>this tutorial</a>
+ * or <a href=https://docs.spring.io/spring-data/jpa/docs/current/reference/html/#repositories>the official documentation</a>.
+ */
+package de.unistuttgart.iste.meitrex.template.persistence.repository;
