+++ src/main/java/de/unistuttgart/iste/meitrex/template/service/TemplateService.java
@@ -1,0 +1,24 @@
+package de.unistuttgart.iste.meitrex.template.service;
+
+import de.unistuttgart.iste.meitrex.generated.dto.Template;
+import de.unistuttgart.iste.meitrex.template.persistence.entity.TemplateEntity;
+import de.unistuttgart.iste.meitrex.template.persistence.mapper.TemplateMapper;
+import de.unistuttgart.iste.meitrex.template.persistence.repository.TemplateRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+@RequiredArgsConstructor
+public class TemplateService {
+
+    private final TemplateRepository templateRepository;
+    private final TemplateMapper templateMapper;
+
+    public List<Template> getAllTemplates() {
+        List<TemplateEntity> templates = templateRepository.findAll();
+        return templates.stream().map(templateMapper::entityToDto).toList();
+    }
+
+}
