+++ src/main/java/de/unistuttgart/iste/meitrex/template/validation/TemplateValidator.java
@@ -1,0 +1,12 @@
+package de.unistuttgart.iste.meitrex.template.validation;
+
+import de.unistuttgart.iste.meitrex.generated.dto.Template;
+import org.springframework.stereotype.Component;
+
+@Component
+public class TemplateValidator {
+
+    public void validateTemplate(Template template) {
+        // add validation logic here
+    }
+}
