+++ src/main/java/de/unistuttgart/iste/meitrex/template/validation/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Put class level validation logic in this package.
+ */
+package de.unistuttgart.iste.meitrex.template.validation;
