--- src/test/java/de/unistuttgart/iste/gits/template/TemplateForMicroservicesApplicationTests.java
@@ -1,13 +1,0 @@
-package de.unistuttgart.iste.gits.template;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class TemplateForMicroservicesApplicationTests {
-
-//	@Test
-//	void contextLoads() {
-//	}
-
-}
