+++ src/test/java/de/unistuttgart/iste/meitrex/template/TemplateForMicroservicesApplicationTests.java
@@ -1,0 +1,12 @@
+package de.unistuttgart.iste.meitrex.template;
+
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class TemplateForMicroservicesApplicationTests {
+
+//	@Test
+//	void contextLoads() {
+//	}
+
+}
